requirements.txt
farm-haystack[opensearch,crawler,pdf,preprocessing,ray]
discord.py
python-dotenv
